Rename array_express_surveyor to array_express.
